requirements.txt
pyautogui>=0.9.54
pywin32>=306
Pillow>=10.0.0
python-dotenv>=1.0.0
